code/RD4AD/requirements.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
imageio==2.23.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-service==2.4.0
networkx==2.8.8
numpy==1.20.3
opencv-python==4.7.0.68
packaging==22.0
pandas==1.5.2
Pillow==8.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.4.1
scikit-image==0.19.3
scikit-learn==1.0
scipy==1.7.1
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.9.1+cu111
torch-summary==1.4.5
torchaudio==0.9.1
torchvision==0.10.1+cu111